Restructure sources into src and include
