+++ include/connection/connection.hpp
@@ -8,7 +8,7 @@
 #include <cstddef>
 #include <netinet/in.h>
 
-#include "v2g_ctx.hpp"
+#include <v2g_ctx.hpp>
 
 /*!
  * \brief initialise TCP/TLS connections
