+++ include/freertos_sync.hpp
@@ -1,7 +1,7 @@
 #ifndef FREERTOS_SYNC_HPP
 #define FREERTOS_SYNC_HPP
 
-#include "freertos_shim.hpp"
+#include <freertos_shim.hpp>
 #include <cerrno>
 #include <ctime>
