+++ include/iso_server.hpp
@@ -5,7 +5,7 @@
 #ifndef ISO_SERVER_HPP
 #define ISO_SERVER_HPP
 
-#include "v2g.hpp"
+#include <v2g.hpp>
 
 struct iso_state {
     const char* description;
