+++ include/v2g_ctx.hpp
@@ -4,7 +4,7 @@
 #ifndef V2G_CTX_H
 #define V2G_CTX_H
 
-#include "v2g.hpp"
+#include <v2g.hpp>
 
 #include <stdbool.h>
